app/service-api/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
asyncpg==0.30.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.10
databricks-sdk==0.38.0
debugpy==1.8.9
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dynaconf==3.2.6
email_validator==2.2.0
evidently==0.5.0
executing==2.1.0
Faker==33.1.0
fastapi==0.115.6
fastapi-cli==0.0.6
fastapi-users==14.0.0
fastapi-users-db-sqlalchemy==6.0.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.1
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl#sha256=7d6ad14cd5078e53147bfbf70fb9d433c6a3865b695fda2657140bbc59a27e29
fsspec==2024.10.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
iterative-telemetry==0.0.9
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
litestar==2.13.0
litestar-htmx==0.4.1
logging==0.4.9.6
makefun==1.15.6
Mako==1.3.7
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.18.0
mlflow-skinny==2.18.0
mpmath==1.3.0
msgspec==0.18.6
multidict==6.1.0
murmurhash==1.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
polyfactory==2.18.1
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwdlib==0.2.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-click==1.8.5
rich-toolkit==0.12.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlmodel==0.0.22
sqlparse==0.5.2
srsly==2.5.0
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.2
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uuid6==2024.7.10
uvicorn==0.32.1
uvloop==0.21.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.0
wcwidth==0.2.13
weasel==0.4.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
